.history/requirements_20250303134141.txt
flask==3.0.0
firebase_admin==6.3.0
Pyrebase4==4.7.1
requests_toolbelt==0.10.1
pylint==3.0.3
flask-cors==4.0.0
pytest==6.2.5
setuptools
pytest-ordering
authlib==1.3.1
curl_cffi==0.9.0
